requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sentence-transformers==5.2.0
numpy==2.4.1
chromadb==1.4.1
openai==2.15.0
python-dotenv==1.2.1
pyyaml==6.0.3
pandas==2.3.3
httpx==0.28.1
pytest==9.0.2
groq==1.0.0
faker==40.1.2
matplotlib
